migration/middleware registration exception

+++ server/migrations/20230210151258_soldier_data.js
@@ -5,13 +5,13 @@
 exports.up = function (knex) {
 	return knex.schema.createTable("soldier_data", function (table) {
 		// table.increments("id");
-		table.integer("DODID", 10).notNullable();
+		table.bigInteger("DODID", 10).notNullable();
 		table.string("SSN", 255).notNullable(); // 000-00-0000
 		table.string("last_name", 255).notNullable();
 		table.string("first_name", 255).notNullable();
 		table.string("middle_initial", 1)
 		table.string("rank", 3)
-		table.integer('company_id');
+		table.bigInteger('company_id');
 		table.foreign('company_id','company_id').references('company_data.id'); // yo mama
 		table.string("mos", 3)
 		table.date("DOB")
@@ -20,7 +20,7 @@ exports.up = function (knex) {
 		table.string("hair_color", 255)
 		table.string("eye_color", 255)
 		table.string("blood_type", 255)
-		table.integer("phone_number", 10)
+		table.bigInteger("phone_number", 10)
 		table.string("address", 255)
 		table.boolean("is_leader").defaultTo(false);
 		table.timestamp('created_at').defaultTo(knex.fn.now());
